requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
djangorestframework_simplejwt==5.4.0
celery==5.4.0
redis==5.2.1
requests==2.32.3